@@ -30,30 +30,31 @@
     "eject": "expo eject"
   },
   "dependencies": {
-    "@react-native-async-storage/async-storage": "^1.14.1",
+    "@react-native-async-storage/async-storage": "~1.17.3",
     "@react-native-community/masked-view": "0.1.10",
     "@react-navigation/native": "5.8.10",
     "@react-navigation/stack": "5.12.8",
     "@unimodules/core": "6.0.0",
     "apisauce": "2.0.0",
-    "expo": "40.0.1",
-    "expo-localization": "9.1.0",
-    "expo-status-bar": "1.0.3",
+    "expo": "45.0.5",
+    "expo-localization": "~13.0.0",
+    "expo-status-bar": "~1.3.0",
     "i18n-js": "3.8.0",
     "mobx": "6.0.4",
     "mobx-react-lite": "3.1.6",
     "mobx-state-tree": "4.0.2",
     "ramda": "0.27.1",
-    "react": "16.13.1",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-40.0.1.tar.gz",
-    "react-native-gesture-handler": "1.8.0",
+    "react": "17.0.2",
+    "react-native": "0.68.2",
+    "react-native-gesture-handler": "~2.2.1",
     "react-native-keychain": "6.2.0",
-    "react-native-safe-area-context": "3.1.9",
-    "react-native-screens": "2.13.0",
+    "react-native-safe-area-context": "4.2.4",
+    "react-native-screens": "~3.11.1",
     "react-native-splash-screen": "3.2.0",
     "react-native-unimodules": "0.11.0",
+    "react-native-uuid": "^2.0.1",
     "reactotron-mst": "3.1.3",
-    "reactotron-react-native": "5.0.0",
+    "reactotron-react-native": "5.0.2",
     "validate.js": "0.13.1"
   },
   "devDependencies": {
