@@ -1,35 +1,25 @@
 import React from "react"
-import { View, Image, ViewStyle, TextStyle, ImageStyle, SafeAreaView } from "react-native"
+import { View, ViewStyle, TextStyle, FlatList } from "react-native"
 import { useNavigation } from "@react-navigation/native"
 import { observer } from "mobx-react-lite"
-import { Button, Header, Screen, Text, Wallpaper } from "../../components"
+import { Button, Screen, Text, Wallpaper } from "../../components"
 import { color, spacing, typography } from "../../theme"
-const bowserLogo = require("./bowser.png")
+import { PrimaryParamList, WelcomeScreenProps } from "../../navigators"
+import { rosterEntries } from "../../services/api"
 
+// #region Styles
 const FULL: ViewStyle = { flex: 1 }
 const CONTAINER: ViewStyle = {
   backgroundColor: color.transparent,
-  paddingHorizontal: spacing[4],
+  padding: spacing[4],
+  justifyContent: "center",
 }
 const TEXT: TextStyle = {
   color: color.palette.white,
   fontFamily: typography.primary,
 }
 const BOLD: TextStyle = { fontWeight: "bold" }
-const HEADER: TextStyle = {
-  paddingTop: spacing[3],
-  paddingBottom: spacing[4] + spacing[1],
-  paddingHorizontal: 0,
-}
-const HEADER_TITLE: TextStyle = {
-  ...TEXT,
-  ...BOLD,
-  fontSize: 12,
-  lineHeight: 15,
-  textAlign: "center",
-  letterSpacing: 1.5,
-}
-const TITLE_WRAPPER: TextStyle = {
+const TITLE_CONTAINER: TextStyle = {
   ...TEXT,
   textAlign: "center",
 }
@@ -40,77 +30,51 @@ const TITLE: TextStyle = {
   lineHeight: 38,
   textAlign: "center",
 }
-const ALMOST: TextStyle = {
-  ...TEXT,
-  ...BOLD,
-  fontSize: 26,
-  fontStyle: "italic",
-}
-const BOWSER: ImageStyle = {
-  alignSelf: "center",
-  marginVertical: spacing[5],
-  maxWidth: "100%",
-}
-const CONTENT: TextStyle = {
-  ...TEXT,
-  color: "#BAB6C8",
-  fontSize: 15,
-  lineHeight: 22,
-  marginBottom: spacing[5],
-}
-const CONTINUE: ViewStyle = {
+const ROSTER_BUTTON: ViewStyle = {
   paddingVertical: spacing[4],
   paddingHorizontal: spacing[4],
+  marginBottom: spacing[4],
   backgroundColor: "#5D2555",
 }
-const CONTINUE_TEXT: TextStyle = {
+const ROSTER_BUTTON_TEXT: TextStyle = {
   ...TEXT,
   ...BOLD,
   fontSize: 13,
   letterSpacing: 2,
 }
-const FOOTER: ViewStyle = { backgroundColor: "#20162D", marginBottom: 64 }
-const FOOTER_CONTENT: ViewStyle = {
-  paddingVertical: spacing[4],
-  paddingHorizontal: spacing[4],
+const ROSTER_CONTAINER: ViewStyle = {
+  marginVertical: spacing[4],
 }
+// #endregion
 
 export const WelcomeScreen = observer(function WelcomeScreen() {
-  const navigation = useNavigation()
-  const nextScreen = () => navigation.navigate("playerList")
+  const navigation = useNavigation<WelcomeScreenProps["navigation"]>()
+
+  const rosters: PrimaryParamList["playerList"][] = rosterEntries
 
   return (
     <View testID="WelcomeScreen" style={FULL}>
       <Wallpaper />
-      <Screen style={CONTAINER} preset="scroll" backgroundColor={color.transparent}>
-        <Header headerTx="welcomeScreen.poweredBy" style={HEADER} titleStyle={HEADER_TITLE} />
-        <Text style={TITLE_WRAPPER}>
-          <Text style={TITLE} text="Your new app, " />
-          <Text style={ALMOST} text="almost" />
-          <Text style={TITLE} text="!" />
-        </Text>
-        <Text style={TITLE} preset="header" tx="welcomeScreen.readyForLaunch" />
-        <Image source={bowserLogo} style={BOWSER} />
-        <Text style={CONTENT}>
-          This probably isn't what your app is going to look like. Unless your designer handed you
-          this screen and, in that case, congrats! You're ready to ship.
-        </Text>
-        <Text style={CONTENT}>
-          For everyone else, this is where you'll see a live preview of your fully functioning app
-          using Ignite.
-        </Text>
-      </Screen>
-      <SafeAreaView style={FOOTER}>
-        <View style={FOOTER_CONTENT}>
-          <Button
-            testID="next-screen-button"
-            style={CONTINUE}
-            textStyle={CONTINUE_TEXT}
-            tx="welcomeScreen.continue"
-            onPress={nextScreen}
+      <Screen style={CONTAINER} backgroundColor={color.transparent}>
+        <View style={TITLE_CONTAINER}>
+          <Text style={TITLE} text="NBA Players" />
+        </View>
+        <View style={ROSTER_CONTAINER}>
+          <FlatList
+            data={rosters}
+            keyExtractor={(item) => item.id}
+            renderItem={({ item }) => (
+              <Button
+                testID={item.id}
+                style={ROSTER_BUTTON}
+                textStyle={ROSTER_BUTTON_TEXT}
+                text={item.name}
+                onPress={() => navigation.navigate("playerList", item)}
+              />
+            )}
           />
         </View>
-      </SafeAreaView>
+      </Screen>
     </View>
   )
 })
