@@ -1,10 +1,7 @@
 import { PlayerModel } from "./player"
 
 test("can be created", () => {
-  const instance = PlayerModel.create({
-    id: 1,
-    name: "Rick Sanchez",
-  })
+  const instance = PlayerModel.create({})
 
   expect(instance).toBeTruthy()
 })
